requirements.txt
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
tqdm==4.67.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
uniface
